requirements.txt
flask
openai
PyYAML